@@ -1,11 +1,16 @@
-namespace MangaLightNovelWebScrape.Src.Websites
+namespace MangaLightNovelWebScrape.Websites
 {
-    public class KinokuniyaUSA
+    public partial class KinokuniyaUSA
     {
-        public static List<string> KinokuniyaUSALinks = new List<string>(); //List of links used to get data from KinokuniyaUSA
-        private static List<string[]> KinokuniyaUSADataList = new List<string[]>(); //List of all the series data from KinokuniyaUSA
-
+        public static List<string> KinokuniyaUSALinks = new(); //List of links used to get data from KinokuniyaUSA
+        private static List<EntryModel> KinokuniyaUSADataList = new(); //List of all the series data from KinokuniyaUSA
         private static readonly NLog.Logger Logger = NLog.LogManager.GetLogger("KinokuniyaUSALogs");
+
+        [GeneratedRegex("(?<=\\d{1,3})[^\\d{1,3}]+.*|,| \\([^()]*\\)|LIGHT NOVEL ")] private static partial Regex ParseTitleNoNumsRegex();
+        [GeneratedRegex("(?<=\\d{1,3}.$)[^\\d{1,3}]+.*|,| \\([^()]*\\)|LIGHT NOVEL ")] private static partial Regex ParseTitleWithNumsRegex();
+        [GeneratedRegex("MANGA", RegexOptions.IgnoreCase, "en-US")] private static partial Regex RemoveMangaFromTitleRegex();
+        [GeneratedRegex("NOVEL", RegexOptions.IgnoreCase, "en-US")] private static partial Regex RemoveNovelFromTitleRegex();
+
         // Manga English Search
         //https://united-states.kinokuniya.com/products?utf8=%E2%9C%93&is_searching=true&restrictBy%5Bavailable_only%5D=1&keywords=world+trigger&taxon=2&x=39&y=4&page=1&per_page=100&form_taxon=109
         // https://united-states.kinokuniya.com/products?utf8=%E2%9C%93&is_searching=true&restrictBy%5Bavailable_only%5D=1&keywords=Skeleton+Knight+in+Another+World&taxon=2&x=39&y=11&page=1&per_page=100
@@ -23,40 +28,40 @@ namespace MangaLightNovelWebScrape.Src.Websites
         public static string TitleParse(string bookTitle, char bookType, string inputTitle)
         {
             string parsedTitle;
-            if (!inputTitle.Any(Char.IsDigit))
+            if (!inputTitle.Any(char.IsDigit))
             {
-                parsedTitle = Regex.Replace(bookTitle.Replace("(Omnibus Edition)", "Omnibus"), @"(?<=\d{1,3})[^\d{1,3}]+.*|,| \([^()]*\)|LIGHT NOVEL ", "");
+                parsedTitle = ParseTitleNoNumsRegex().Replace(bookTitle.Replace("(Omnibus Edition)", "Omnibus").Replace("Volume", "Vol"), "");
             }
             else
             {
-                parsedTitle = Regex.Replace(bookTitle.Replace("(Omnibus Edition)", "Omnibus"), @"(?<=\d{1,3}.$)[^\d{1,3}]+.*|,| \([^()]*\)|LIGHT NOVEL ", "");
+                parsedTitle = ParseTitleWithNumsRegex().Replace(bookTitle.Replace("(Omnibus Edition)", "Omnibus").Replace("Volume", "Vol"), "");
             }
 
             // Check to see if after parsing, the volume number is still there if not then fix/add it back
-            if (!MasterScrape.Similar(parsedTitle, inputTitle))
+            if (!EntryModel.Similar(parsedTitle, inputTitle))
             {
                 parsedTitle = Regex.Replace(new Regex(@"\((.*?)\)").Match(bookTitle).Groups[0].ToString(), @"\(|\)", "");
-                if (!Char.IsDigit(parsedTitle, parsedTitle.Length - 1))
+                if (!char.IsDigit(parsedTitle, parsedTitle.Length - 1))
                 {
-                    parsedTitle += $" {new Regex(@"(\d+)").Match(bookTitle).Groups[0].ToString()}";
+                    parsedTitle += $" {new Regex(@"(\d+)").Match(bookTitle).Groups[0]}";
                 }
                 // Console.WriteLine("Fixed Title 1 = " + parsedTitle);
             }
 
             if (parsedTitle.Contains("Manga", StringComparison.OrdinalIgnoreCase))
             {
-                parsedTitle = Regex.Replace(parsedTitle, "MANGA", "", RegexOptions.IgnoreCase);
+                parsedTitle = RemoveMangaFromTitleRegex().Replace(parsedTitle, "");
             }
             else if (parsedTitle.Contains("Novel", StringComparison.OrdinalIgnoreCase))
             {
-                parsedTitle = Regex.Replace(parsedTitle, "NOVEL", "", RegexOptions.IgnoreCase);
+                parsedTitle = RemoveNovelFromTitleRegex().Replace(parsedTitle, "");
                 // Console.WriteLine("Fixed Title 2 = " + parsedTitle);
             }
             
 
             if (!parsedTitle.Contains("Vol", StringComparison.OrdinalIgnoreCase) && !parsedTitle.Contains("Box Set", StringComparison.OrdinalIgnoreCase))
             {
-                parsedTitle = parsedTitle.Insert(Regex.Match(parsedTitle, @"\d{1,3}").Index, "Vol ");
+                parsedTitle = parsedTitle.Insert(MasterScrape.FindVolNumRegex().Match(parsedTitle).Index, "Vol ");
                 // Console.WriteLine("Fixed Title 3 = " + parsedTitle);
             }
             else if (!parsedTitle.Contains("Box Set", StringComparison.OrdinalIgnoreCase))
@@ -69,22 +74,22 @@ namespace MangaLightNovelWebScrape.Src.Websites
             {
                 parsedTitle = parsedTitle.Insert(parsedTitle.IndexOf("Vol"), "Novel ");
             }
-            else if (bookType == 'M' && !parsedTitle.Contains("Box Set", StringComparison.OrdinalIgnoreCase) && !MasterScrape.Similar(parsedTitle.Substring(0, parsedTitle.IndexOf("Vol") - 1), inputTitle))
+            else if (bookType == 'M' && !parsedTitle.Contains("Box Set", StringComparison.OrdinalIgnoreCase) && !EntryModel.Similar(parsedTitle[..(parsedTitle.IndexOf("Vol") - 1)], inputTitle))
             {
-                parsedTitle = parsedTitle.Remove(0, parsedTitle.IndexOf("Vol")).Insert(0, char.ToUpper(inputTitle[0]) + inputTitle.ToLower().Substring(1) + " ");
+                parsedTitle = parsedTitle.Remove(0, parsedTitle.IndexOf("Vol")).Insert(0, char.ToUpper(inputTitle[0]) + inputTitle.ToLower()[1..] + " ");
             }
 
             return parsedTitle.Trim();
         }
 
-        public static List<string[]> GetKinokuniyaUSAData(string bookTitle, char bookType, bool memberStatus, byte currPageNum, EdgeOptions edgeOptions)
+        public static List<EntryModel> GetKinokuniyaUSAData(string bookTitle, char bookType, bool memberStatus, byte currPageNum, EdgeOptions edgeOptions)
         {
-            WebDriver dummyDriver = new EdgeDriver(@"DriverExecutables/Edge", edgeOptions);
+            WebDriver dummyDriver = new EdgeDriver(edgeOptions);
             edgeOptions.AddArgument("user-agent=" + dummyDriver.ExecuteScript("return navigator.userAgent").ToString().Replace("Headless", ""));
             dummyDriver.Quit();
 
-            EdgeDriver edgeDriver = new EdgeDriver(@"DriverExecutables/Edge", edgeOptions);
-            WebDriverWait wait = new WebDriverWait(edgeDriver, TimeSpan.FromSeconds(30));
+            EdgeDriver edgeDriver = new(edgeOptions);
+            WebDriverWait wait = new(edgeDriver, TimeSpan.FromSeconds(30));
 
             try
             {
@@ -96,13 +101,14 @@ namespace MangaLightNovelWebScrape.Src.Websites
                         // Click the Manga button so it only shows manga and wait for DOM to fully load
                         wait.Until(driver => driver.FindElement(By.XPath("//div[@id='loading' and @style='display: none;']")));
                         edgeDriver.ExecuteScript("arguments[0].click();", wait.Until(driver => driver.FindElement(By.XPath("//p[contains(text(), 'English Books')]/following-sibling::ul//a[contains(text(), 'Manga')]"))));
+                        Logger.Debug("Clicked Manga Button");
                     }
                     wait.Until(driver => driver.FindElement(By.XPath("//div[@id='loading' and @style='display: none;']")));
                     wait.Until(driver => driver.FindElement(By.XPath("//ul[@class='sortMenu']//li//a[contains(text(), 'List')]"))).Click();
                     wait.Until(driver => driver.FindElement(By.XPath("//div[@id='loading' and @style='display: none;']")));
 
                     // Initialize the html doc for crawling
-                    HtmlDocument doc = new HtmlDocument();
+                    HtmlDocument doc = new();
                     doc.LoadHtml(edgeDriver.PageSource);
 
                     // Get the page data from the HTML doc
@@ -113,7 +119,10 @@ namespace MangaLightNovelWebScrape.Src.Websites
                     
                     // for (int x = 1; x < priceData.Count; x+=2)
                     // {
-                    //     Logger.Debug($"[{titleData[x / 2].InnerText}, {priceData[x - 1].InnerText.Trim()}, {priceData[x ].InnerText.Trim()}, {stockStatusData[x / 2].InnerText}, KinokuniyaUSA");
+                    //     if(MasterScrape.RemoveNonWordsRegex().Replace(titleData[x / 2].InnerText, "").Contains(MasterScrape.RemoveNonWordsRegex().Replace(bookTitle, ""), StringComparison.OrdinalIgnoreCase))
+                    //     {
+                    //         Logger.Debug($"[{titleData[x / 2].InnerText}, {priceData[x - 1].InnerText.Trim()}, {priceData[x].InnerText.Trim()}, {stockStatusData[x / 2].InnerText}, KinokuniyaUSA");
+                    //     }
                     // }
 
                     // Remove all of the novels from the list if user is searching for manga
@@ -121,7 +130,7 @@ namespace MangaLightNovelWebScrape.Src.Websites
                     {
                         for (int x = 1; x < priceData.Count; x+=2)
                         {
-                            if (titleData[x / 2].InnerText.Contains("Novel", StringComparison.OrdinalIgnoreCase) && bookType == 'M' || !titleData[x / 2].InnerText.Any(Char.IsDigit) || titleData[x / 2].InnerText.Contains("Volume", StringComparison.OrdinalIgnoreCase))
+                            if (titleData[x / 2].InnerText.Contains("Novel", StringComparison.OrdinalIgnoreCase) && bookType == 'M' || !titleData[x / 2].InnerText.Any(char.IsDigit))
                             {
                                 titleData.RemoveAt(x / 2);
                                 stockStatusData.RemoveAt(x / 2);
@@ -137,9 +146,10 @@ namespace MangaLightNovelWebScrape.Src.Websites
                     }
 
                     string stockStatus = "";
-                    if (memberStatus) // If the user is a member only take the discounted prices (10%)
+                    int pos = memberStatus ? 1 : 0; // Position to start in array
+                    for (int x = pos; x < priceData.Count; x+=2)
                     {
-                        for (int x = 1; x < priceData.Count; x+=2)
+                        if (MasterScrape.RemoveNonWordsRegex().Replace(titleData[x / 2].InnerText, "").Contains(MasterScrape.RemoveNonWordsRegex().Replace(bookTitle, ""), StringComparison.OrdinalIgnoreCase))
                         {
                             if (stockStatusData[x / 2].InnerText.Contains("Out of stock"))
                             {
@@ -157,38 +167,11 @@ namespace MangaLightNovelWebScrape.Src.Websites
                             // if (titleData[x / 2].InnerText.Contains("3"))
                             // {
                             //     Logger.Debug("Check" + TitleParse(titleData[x / 2].InnerText, bookType, bookTitle));
-                            //     KinokuniyaUSADataList.Add(new string[]{TitleParse(titleData[x / 2].InnerText, bookType, bookTitle), "$4.00", stockStatus, "KinokuniyaUSA"});
+                            //     KinokuniyaUSADataList.Add(new EntryModel(TitleParse(titleData[x / 2].InnerText, bookType, bookTitle), "$4.00", stockStatus, "KinokuniyaUSA"));
                             //     continue;
                             // }
 
-                            KinokuniyaUSADataList.Add(new string[]{TitleParse(titleData[x / 2].InnerText, bookType, bookTitle), priceData[x].InnerText.Trim(), stockStatus, "KinokuniyaUSA"});
-                        }
-                    }
-                    else
-                    {
-                        for (int x = 0; x < priceData.Count; x+=2)
-                        {
-                            if (stockStatusData[x / 2].InnerText.Contains("Out of stock"))
-                            {
-                                stockStatus = "OOS";
-                            }
-                            else if (stockStatusData[x / 2].InnerText.Contains("Pre Order"))
-                            {
-                                stockStatus = "PO";
-                            }
-                            else if (stockStatusData[x / 2].InnerText.Contains("In stock"))
-                            {
-                                stockStatus = "IS";
-                            }
-                            
-                            // if (titleData[x / 2].InnerText.Contains("3"))
-                            // {
-                            //    Logger.Debug("Check" + TitleParse(titleData[x / 2].InnerText, bookType, bookTitle));
-                            //     KinokuniyaUSADataList.Add(new string[]{TitleParse(titleData[x / 2].InnerText, bookType, bookTitle), "$4.00", stockStatus, "KinokuniyaUSA"});
-                            //     continue;
-                            // }
-
-                            KinokuniyaUSADataList.Add(new string[]{TitleParse(titleData[x / 2].InnerText, bookType, bookTitle), priceData[x].InnerText.Trim(), stockStatus, "KinokuniyaUSA"});
+                            KinokuniyaUSADataList.Add(new EntryModel(TitleParse(titleData[x / 2].InnerText, bookType, bookTitle), priceData[x].InnerText.Trim(), stockStatus, "KinokuniyaUSA"));
                         }
                     }
 
@@ -210,20 +193,23 @@ namespace MangaLightNovelWebScrape.Src.Websites
             }
 
             //Print data to a txt file
-            using (StreamWriter outputFile = new StreamWriter(@"Data\KinokuniyaUSAData.txt"))
+            if (MasterScrape.IsDebugEnabled)
             {
-                if (KinokuniyaUSADataList.Count != 0)
+                using (StreamWriter outputFile = new(@"Data\KinokuniyaUSAData.txt"))
                 {
-                    foreach (string[] data in KinokuniyaUSADataList)
+                    if (KinokuniyaUSADataList.Count != 0)
                     {
-                        outputFile.WriteLine("[" + data[0] + ", " + data[1] + ", " + data[2] + ", " + data[3] + "]");
-                        Logger.Debug("[" + data[0] + ", " + data[1] + ", " + data[2] + ", " + data[3] + "]");
+                        foreach (EntryModel data in KinokuniyaUSADataList)
+                        {
+                            outputFile.WriteLine(data.ToString());
+                            Logger.Debug(data.ToString());
+                        }
                     }
-                }
-                else{
-                    outputFile.WriteLine(bookTitle + " Does Not Exist at KinokuniyaUSA");
-                }
-            } 
+                    else{
+                        outputFile.WriteLine(bookTitle + " Does Not Exist at KinokuniyaUSA");
+                    }
+                } 
+            }
             return KinokuniyaUSADataList;
         }
     }
